requirements.txt
requests
PyYAML
PySocks
colorama
